requirements.txt
# Basic Python dependencies
requests
pandas
numpy
matplotlib
seaborn
pyyaml
xlsxwriter